Revert to using bash
fish shell doesn't play nicely with stuff I rely on in vim (yank etc)
and it's far from obvious how I get it to work (lots of googling, and
chatgpt'ing). I kind of got it working in Linux, but cannot get it to
work on macos (work device). Fish is nice, but not so life-changing that
I can't go back to bash. May play around with it in the future.