Revert back to fish shell
I failed to spot that I was not using the right clipboard for mac in my
vimrc. Fixed that by setting it to unnamed (as opposed to unnamedplus as
it is on linux) and I can yank and paste in and out of Vim on fish no
issue at all.